Fix running Hardware tests (#531)
When running HW tests with

RUN_DEVICE_TESTS=true yarn test

the HW tests all failed with:

Error: Resolution method is overspecified. Specify a callback *or* return a Promise; not both.

Basically the test framework no longer allows to return a Promise and to
end with a callback. As the callback was only used with the return
promise anyhow, I simplified the code by dropping the done callback,
just using the promise.

Signed-off-by: Daniel Friederich <d.friederich@samsung.com>